Adds sampler test and renamed PercentageSampler namespace.
